move out vms
